Extracted buffer test class to separate compilation unit for reuse
